+++ src/api/controllers/cart/index.js
@@ -1,10 +1,13 @@
 const Express = require('express');
 
+const etupay = require('../../utils/etupay.js');
+
 const CreateCart = require('./create.js');
 const DeleteItemFromCart = require('./delete.js');
 const { AddItemToCart, CheckAddItem } = require('./add-item.js');
 const { Edit, CheckEdit } = require('./edit.js');
 const GetItemFromCart = require('./get-item-from-cart.js');
+const SuccessfulPayment = require('./etupay.callback');
 
 const cartId = 'cartId';
 const itemId = 'itemId';
@@ -37,6 +40,7 @@ const Cart = models => {
             models.Attribute
         )
     );
+    router.get('/return', etupay.middleware, SuccessfulPayment());
     return router;
 };
